+++ src/components/Icon/index.tsx
@@ -1,2 +1,3 @@
 export { default } from "./Icon";
+export type { IconSize } from "./Icon";
 export type { Props as IconProps, IconName } from "./Icon";